requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
numpy>=1.24.0
google-generativeai>=0.3.0
plotly>=5.18.0
matplotlib>=3.8.0
seaborn>=0.13.0
scikit-learn>=1.3.0  # for OLS trendline in plotly
rich>=13.7.0
Flask>=3.0.0
Flask-Session>=0.5.0
python-dotenv>=1.0.0
statsmodels>=0.14.0  # for additional statistical analysis
scipy>=1.11.0  # for statistical computations